# contracts/catalog-api/internal-insights/v1/openapi.yaml
---
openapi: 3.1.0
info:
  title: GrooveMap internal insights API
  version: 1.1.0
  description: Catalog API-owned contract consumed by the analytics engine.
security:
  - InternalSecret: []
paths:
  /api/internal/insights/artist-centrality:
    get:
      operationId: artist_centrality
      parameters:
        - $ref: "#/components/parameters/ArtistCentralityLimit"
      responses:
        "200": {$ref: "#/components/responses/Items"}
  /api/internal/insights/genre-trends:
    get:
      operationId: genre_trends
      responses:
        "200": {$ref: "#/components/responses/Items"}
  /api/internal/insights/label-longevity:
    get:
      operationId: label_longevity
      parameters:
        - $ref: "#/components/parameters/LabelLongevityLimit"
      responses:
        "200": {$ref: "#/components/responses/Items"}
  /api/internal/insights/anniversaries:
    get:
      operationId: anniversaries
      parameters:
        - $ref: "#/components/parameters/AnniversaryYear"
        - $ref: "#/components/parameters/AnniversaryMonth"
        - $ref: "#/components/parameters/AnniversaryMilestones"
      responses:
        "200": {$ref: "#/components/responses/Items"}
  /api/internal/insights/data-completeness:
    get:
      operationId: data_completeness
      responses:
        "200": {$ref: "#/components/responses/Items"}
  /api/internal/insights/community-enrichment:
    get:
      operationId: community_enrichment
      x-groovemap-max-processing-seconds: 1500
      responses:
        "200": {$ref: "#/components/responses/CommunityEnrichment"}
  /api/internal/insights/rarity-scores:
    get:
      operationId: rarity_scores
      responses:
        "200": {$ref: "#/components/responses/RarityScores"}
components:
  securitySchemes:
    InternalSecret:
      type: apiKey
      in: header
      name: X-Internal-Secret
  parameters:
    ArtistCentralityLimit:
      name: limit
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 500, default: 100}
    LabelLongevityLimit:
      name: limit
      in: query
      required: false
      schema: {type: integer, minimum: 1, maximum: 500, default: 50}
    AnniversaryYear:
      name: year
      in: query
      required: true
      schema: {type: integer}
    AnniversaryMonth:
      name: month
      in: query
      required: true
      schema: {type: integer, minimum: 1, maximum: 12}
    AnniversaryMilestones:
      name: milestones
      in: query
      required: false
      schema: {type: string, default: "25,30,40,50,75,100"}
  schemas:
    CommunityEnrichmentResult:
      type: object
      required: [enriched, skipped, errors, remaining]
      properties:
        enriched: {type: integer, minimum: 0}
        skipped: {type: integer, minimum: 0}
        errors: {type: integer, minimum: 0}
        remaining: {type: integer, minimum: 0}
        error:
          type: string
          enum: [no_credentials]
      additionalProperties: false
    RarityScoreItem:
      description: >-
        One release's rarity computation. Per ADR 0007 the core score is media-neutral;
        `pressing_scarcity` is populated only for grooved media (vinyl, shellac,
        grooved_other) and is `null` otherwise. `format_rarity` is deprecated for one
        minor version: still computed and present at weight `0.0`, superseded by
        `medium_rarity`.
      type: object
      required:
        - release_id
        - title
        - artist_name
        - year
        - rarity_score
        - tier
        - hidden_gem_score
        - pressing_scarcity
        - label_catalog
        - format_rarity
        - temporal_scarcity
        - graph_isolation
        - collection_prevalence
        - medium_rarity
        - media_families
        - family_signals
      properties:
        release_id: {type: string}
        title: {type: string}
        artist_name: {type: string}
        year: {type: [integer, "null"]}
        rarity_score: {type: number}
        tier: {type: string}
        hidden_gem_score: {type: number}
        pressing_scarcity:
          type: [number, "null"]
          description: Grooved-only signal; `null` when no family extension claims the release.
        label_catalog: {type: number}
        format_rarity:
          type: number
          deprecated: true
          description: Raw Discogs format name signal, kept for one minor version at weight 0.0.
        temporal_scarcity: {type: number}
        graph_isolation: {type: number}
        collection_prevalence: {type: number}
        medium_rarity:
          type: [number, "null"]
          description: Canonical-medium rarity score (ADR 0007); replaces `format_rarity`.
        media_families:
          type: array
          items: {type: string}
          description: Canonical ADR 0007 media family ids the release covers.
        family_signals:
          type: object
          additionalProperties: true
          description: Per-family-extension signal scores keyed by module id (e.g. `grooved`).
      additionalProperties: true
  responses:
    Items:
      description: A successful computation input response.
      content:
        application/json:
          schema:
            type: object
            required: [items]
            properties:
              items:
                type: array
                items: {type: object, additionalProperties: true}
            additionalProperties: false
    CommunityEnrichment:
      description: Community enrichment counters for the completed bounded batch.
      content:
        application/json:
          schema: {$ref: "#/components/schemas/CommunityEnrichmentResult"}
    RarityScores:
      description: Rarity score data for every release, media-neutral core plus family extensions.
      content:
        application/json:
          schema:
            type: object
            required: [items]
            properties:
              items:
                type: array
                items: {$ref: "#/components/schemas/RarityScoreItem"}
            additionalProperties: false